# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Туристический ассистент API",
    "description": "REST API веб-интерфейса: multi-user SaaS, поездки в SQLite, асинхронная сборка программы LangGraph, BYOK OpenRouter.",
    "version": "2.0.0"
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "operationId": "me_api_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/profile": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Get Profile",
        "operationId": "get_profile_api_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/profile/settings": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Get Settings",
        "operationId": "get_settings_api_profile_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "profile"
        ],
        "summary": "Update Settings",
        "operationId": "update_settings_api_profile_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/profile/settings/llm-key": {
      "delete": {
        "tags": [
          "profile"
        ],
        "summary": "Delete Llm Key",
        "operationId": "delete_llm_key_api_profile_settings_llm_key_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "user"
        ],
        "title": "AuthResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "minLength": 3,
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "ProfileResponse": {
        "properties": {
          "preferences": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TripPreferences"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "preferences"
        ],
        "title": "ProfileResponse"
      },
...
      "RegisterRequest": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "minLength": 3,
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "RegisterRequest"
      },
...
      "SettingsResponse": {
        "properties": {
          "llm_key_configured": {
            "type": "boolean",
            "title": "Llm Key Configured"
          },
          "llm_key_preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Key Preview"
          },
          "llm_base_url": {
            "type": "string",
            "title": "Llm Base Url"
          },
          "llm_model": {
            "type": "string",
            "title": "Llm Model"
          }
        },
        "type": "object",
        "required": [
          "llm_key_configured",
          "llm_base_url",
          "llm_model"
        ],
        "title": "SettingsResponse"
      },
...
      "TripPreferences": {
        "properties": {
          "pace": {
            "type": "string",
            "enum": [
              "relaxed",
              "moderate",
              "packed"
            ],
            "title": "Pace",
            "description": "Темп поездки (в UI скрыт, по умолчанию packed)"
          },
          "budget": {
            "type": "string",
            "enum": [
              "economy",
              "medium",
              "unlimited"
            ],
            "title": "Budget",
            "description": "Legacy: не используется в поиске"
          },
          "interests": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Interests",
            "description": "Legacy: не влияет на пул POI"
          },
          "cuisine": {
            "type": "string",
            "title": "Cuisine",
            "description": "Legacy: поиск ресторанов отключён",
            "default": ""
          },
          "min_restaurant_rating": {
            "type": "number",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Min Restaurant Rating",
            "description": "Legacy: не используется",
            "default": 4.0
          },
          "transport_preference": {
            "type": "string",
            "enum": [
              "metro",
              "taxi",
              "walking",
              "mixed"
            ],
            "title": "Transport Preference",
            "description": "Передвижение (в UI скрыт, по умолчанию mixed)"
          },
          "travel_party": {
            "type": "string",
            "enum": [
              "solo",
              "couple",
              "family",
              "friends",
              "parent_child",
              "family_two"
            ],
            "title": "Travel Party",
            "description": "Состав группы"
          },
          "special_notes": {
            "type": "string",
            "title": "Special Notes",
            "description": "Legacy: доп. пожелания",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "pace",
          "budget",
          "transport_preference",
          "travel_party"
        ],
        "title": "TripPreferences",
        "description": "Результат опросника перед планированием программы."
      },
...
      "UpdateSettingsRequest": {
        "properties": {
          "llm_api_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Api Key"
          },
          "llm_base_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Base Url"
          },
          "llm_model": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Model"
          }
        },
        "type": "object",
        "title": "UpdateSettingsRequest"
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}